+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [] - ?
+## [1.2.0] - 2019-10-29
 
 ### Added
 - Introduced SCORE_THRESHOLD parameter as a minimum patient score threshold for returned results
@@ -7,16 +7,19 @@
 - Command line returns app version
 
 
+
 ## [1.1.1] - 2019-04-25
 
 ### Modified
 -  Fixed bug in phenotype matching when no OMIM or no HPO terms are available
 
 
+
 ## [1.1.0] - 2019-04-25
 
 ### Modified
 -  patient-similarity against all patients in database if query patient has HPO term
+
 
 
 ## [1.0.0] - 2019-04-18